SCHEMA: Use strict types, avoid cluttering docs
The demo showed that 1 in `Model.X` was being interpreted as "1", so
adjust the ordering of the types in the Union, and change to strict
types to avoid similar problems. This makes the docs ugly to no benefit,
so use a hack to override these types only during documentation. I've
tested and this check is not activated when called from a jupytext cell,
so we will not lose the strict types.